+++ scraper-yahoo/src/main/java/org/mirror/DTO/HistoricalQuote.java
@@ -2,16 +2,17 @@ package org.mirror.DTO;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 
 @Data
 public class HistoricalQuote {
     private String symbol;
     private Calendar date;
-    private Double high;
-    private Double low;
-    private Double open;
-    private Double close;
+    private BigDecimal high;
+    private BigDecimal low;
+    private BigDecimal open;
+    private BigDecimal close;
     private Long volume;
-    private Double adjClose;
+    private BigDecimal adjClose;
 }
